==> 02_feature_engineering.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

df = pd.read_csv('data_agg/final_aggregated_output_no_outlier.csv', index_col=0)
df

def calc_means(df, cols, activity):
    means = {}
    for c in cols:
        means[c] = df[df["Activity"] == activity][c].mean()
    return means

calc_means(df, ["Acceleration x (m/s^2)", "Acceleration y (m/s^2)", "Acceleration z (m/s^2)"], "Sport")

[col for col in df.columns if "mean" in col]

from ML4QS.Python3Code.Chapter4.TemporalAbstraction import NumericalAbstraction

num_abs = NumericalAbstraction()

def calc_means(df, cols, method="slope"):
    means = {}
    for c in cols:
        means[c] = num_abs.aggregate_value(df[c].abs(), 10, method)
    return means

calc_means(df, ["Linear Acceleration x (m/s^2)", "Linear Acceleration y (m/s^2)", "Linear Acceleration z (m/s^2)"], "slope")
